Moved game objects to their own file
