+++ game/base.py
@@ -1,0 +1,42 @@
+import yaml
+
+class GameObject(yaml.YAMLObject):
+    __metaclass__ = yaml.YAMLObjectMetaclass
+
+    def __init__(self):
+        super(GameObject,self).__init__()
+        # Call byteify to make sure all unicode variables are saved as strings
+        # This makes it easier to save in yaml
+        self.__dict__ = self.byteify(self.__dict__)
+
+    def byteify(self, data):
+        """Make sure everything in data is a string, not unicode (for nicer yaml)."""
+        if isinstance(data, dict):
+            result = {}
+            for key, value in data.iteritems():
+                if isinstance(key, unicode):
+                    key = key.encode('utf-8')
+                if isinstance(value, unicode):
+                    value = value.encode('utf-8')
+                if isinstance(value, dict) or isinstance(value, list):
+                    value = self.byteify(value)
+                result[key] = value
+        elif isinstance(data, list):
+            result = []
+            for item in data:
+                if isinstance(item, unicode):
+                    item = item.encode('utf-8')
+                if isinstance(item, dict) or isinstance(item, list):
+                    item = self.byteify(value)
+        else:
+            result = None
+
+        return result
+
+    def __repr__(self):
+        return str(self.__dict__)
+
+    @classmethod
+    def from_yaml(cls, loader, node):
+        fields = loader.construct_mapping(node, deep=True)
+        return cls(**fields)
