+++ game/sector.py
@@ -1,0 +1,3 @@
+from base import GameObject
+
+class Sector(GameObject):
