+++ game/ship.py
@@ -1,0 +1,23 @@
+import uuid
+from base import GameObject
+
+class Ship(GameObject):
+    yaml_tag = "!Ship"
+
+    def __init__(
+        self,
+        name = None,
+        id = str(uuid.uuid4())
+    ):
+        self.name = name
+        self.id = id
+
+        # Parent init should be called at end of __init__
+        super(Ship,self).__init__()
+
+    def __str__(self):
+        return "%s (name=%s, id=%s)" % (
+            self.__class__.__name__,
+            self.name,
+            self.id,
+        )
